moved events specifics, configuration
