+++ src/main/kotlin/org/developerden/codosseum/server/routes/event/EventsRoute.kt
@@ -1,4 +1,4 @@
-package org.developerden.codosseum.server.routes
+package org.developerden.codosseum.server.routes.event
 
 import io.github.tabilzad.ktor.annotations.GenerateOpenApi
 import io.github.tabilzad.ktor.annotations.KtorDescription
@@ -10,7 +10,6 @@ import kotlinx.coroutines.flow.collectLatest
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
-import org.developerden.codosseum.EventBus
 import org.developerden.codosseum.server.Events
 import org.developerden.codosseum.server.koin.inject
 import org.developerden.codosseum.validation.FailedTest
