+++ src/main/kotlin/org/developerden/codosseum/validation/SolutionValidation.kt
@@ -8,18 +8,18 @@ import kotlinx.coroutines.flow.toSet
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import kotlinx.serialization.Serializable
-import org.developerden.codosseum.EventBus
+import org.developerden.codosseum.server.routes.event.EventBus
 import org.developerden.codosseum.challenge.Challenge
 import org.developerden.codosseum.sandkasten.api.apis.ProgramsApi
 import org.developerden.codosseum.sandkasten.api.models.BuildRequest
 import org.developerden.codosseum.sandkasten.api.models.BuildRequestMainFile
 import org.developerden.codosseum.sandkasten.api.models.RunRequest
-import org.developerden.codosseum.server.routes.TestCompleteEvent
+import org.developerden.codosseum.server.routes.event.TestCompleteEvent
 
 suspend fun validateSolutions(
-	eventsBus: EventBus,
-	api: ProgramsApi,
-	challenge: Challenge
+  eventsBus: EventBus,
+  api: ProgramsApi,
+  challenge: Challenge
 ): SolutionValidationResult {
 	val compiledSolution = api.compile(
 		BuildRequest(
